cache versionstamp
